Rename `concurrency` directory to `backends`
Closes #659
