+++ httpx/client.py
@@ -6,7 +6,7 @@ from types import TracebackType
 import hstspreload
 
 from .auth import Auth, AuthTypes, BasicAuth, FunctionAuth
-from .concurrency.base import ConcurrencyBackend
+from .backends.base import ConcurrencyBackend
 from .config import (
     DEFAULT_MAX_REDIRECTS,
     DEFAULT_POOL_LIMITS,
